+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### 0.6-pre1
 
+- **virustotal**: Command now works again with V3 API
+    - Requires Api key for all operations
 - **decompress**: Add `Expand-ZstdArchive` function for extracting standalone zstd archives
 - **scoop-install**: Allow modules to be installed globally
 - **scoop-install**: Prevent repeated installation of same manifest/url/local file